test(web-components): remove text area test file that causes timeout errors
Remove text area test file that causes timeout errors - replace with one test with Cypress test and move the other tests into text field test file
